Adjusted plugin install directory
This will make it simpler to have install plugins and not have them
tracked by git, while still keeping the configs of the plugins be
stored in git. This was the easiest solution I found, the .gitignore
file was difficult to set up so that it ignores all but one folder.