Changed nvim terminal key binding
Had to change this one, because it was causing troubles for examples
when I was committing something via the nvim terminal. In that case a
new nvim (inside nvim) was opened and the ESC key put me back into the
terminal buffer which I couldn't modify when I wanted to save the
commit messages with ':w'.

The ESC key was basically mapped to server two different functions in
two different contexts. However when these overlapped I was unable to
use the terminal functionality in nvim properly.