Added function to create a new temporary file
This is for convenience, since having a file saved has some benefits
over an unnamed buffer. This also helps to work around some issues that
come up with using the 'MaximizeToggle' function which creates a new
unnamed buffer when I return from the maximized buffer back to the
original layout. In such a case, I always have to switch back to the
unnamed buffer that I created on purpose. This doesn't happen when I
have a temporary file.