Added key binding to open man page under cursor
This is helpful, as vim is used to view man pages bu default. With this
shortcut, I can go further into other man pages which are referred to in
the currently open one.
Note: There is still an improvement to make, namely since the buffer
which is opened the first time the 'man' command is issued sets 'nomodifiable'
the buffer disappears when I come back to it from a 'MaximizeToggle'.